requirements.txt
google-api-python-client==2.110.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
pillow>=10.0.0
python-dotenv==1.0.0
apscheduler==3.10.4
requests==2.31.0
tqdm==4.66.1
tenacity==8.2.3
pytz==2023.3 